+++ src/main/java/kitchenpos/menu/repository/MenuProductRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.menu.repository;
 
 import java.util.List;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
