+++ src/main/java/kitchenpos/menu/repository/MenuRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.menu.repository;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
