+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -5,15 +5,16 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
-import kitchenpos.application.request.MenuCreateRequest;
-import kitchenpos.application.dto.ProductQuantityDto;
-import kitchenpos.application.response.MenuResponse;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.Product;
+import kitchenpos.menu.application.request.MenuCreateRequest;
+import kitchenpos.dto.ProductQuantityDto;
+import kitchenpos.menu.application.response.MenuResponse;
+import kitchenpos.menu.application.MenuService;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.product.domain.Product;
 import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.fixture.ProductFixture;
-import kitchenpos.repository.MenuGroupRepository;
-import kitchenpos.repository.ProductRepository;
+import kitchenpos.menu.repository.MenuGroupRepository;
+import kitchenpos.product.repository.ProductRepository;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
