+++ src/test/java/kitchenpos/domain/vo/MenuGroupNameTest.java
@@ -3,6 +3,7 @@ package kitchenpos.domain.vo;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import kitchenpos.dto.vo.MenuGroupName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.params.ParameterizedTest;
