+++ src/test/java/kitchenpos/fixture/MenuFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
 
 @SuppressWarnings("NonAsciiCharacters")
 public class MenuFixture {
