+++ src/test/java/kitchenpos/fixture/MenuGroupFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 @SuppressWarnings("NonAsciiCharacters")
 public class MenuGroupFixture {
